fixing rendering of mdx files at build time
