+++ src/quickstart/00-main.mdx
@@ -1,0 +1,32 @@
+<script>
+  {/* todo fix the prettier formatting  */}
+  /* import all the mdx files into the main mdx file. See mdsvex example here: https://mdsvex.pngwn.io/playground
+  */ import Introduction from '../quickstart/01-introduction.mdx'; import Unboxing
+  from '../quickstart/02-unboxing-your-airnote.mdx'; import Setup from '../quickstart/03-airnote-setup.mdx';
+  import Dashboard from '../quickstart/04-visit-your-device-dashboard.mdx'; import
+  AirQuality from '../quickstart/05-understanding-air-quality.mdx'; import FAQ from
+  '../quickstart/06-faq.mdx';
+</script>
+
+# Airnote Quickstart
+
+- [Introduction](#introduction)
+- [Unboxing Your Airnote](#unboxing-your-airnote)
+- [Airnote Setup](#airnote-setup)
+- [Visit Your Device Dashboard](#visit-your-device-dashboard)
+- [Understanding Air Quality](#understanding-air-quality)
+- [FAQ](#faq)
+
+<hr />
+
+<Introduction />
+
+<Unboxing />
+
+<Setup />
+
+<Dashboard />
+
+<AirQuality />
+
+<FAQ />
