+++ src/quickstart/02-unboxing-your-airnote.mdx
