+++ src/quickstart/03-airnote-setup.mdx
