+++ src/quickstart/06-faq.mdx
