--- src/setup-files/00 Main.mdx
@@ -1,10 +1,0 @@
-## Airnote Quickstart
-
-- [Introduction](#introduction)
-- [Unboxing Your Airnote](#unboxing-your-airnote)
-- [Airnote Setup](#airnote-setup)
-- [Visit Your Device Dashboard](#visit-your-device-dashboard)
-- [Understanding Air Quality](#understanding-air-quality)
-- [FAQ](#faq)
-
-<hr />
