+++ static/_redirects
@@ -1,1 +1,1 @@
-https://start.airnote.live https://airnote.live/setup-guide 301!
+https://start.airnote.live https://airnote.live/quickstart 301!
